+++ src/main/java/roomescape/reservation/controller/ReservationController.java
@@ -1,4 +1,4 @@
-package roomescape.reservation;
+package roomescape.reservation.controller;
 
 import jakarta.validation.Valid;
 import java.util.List;
@@ -9,6 +9,10 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import roomescape.reservation.entity.Reservation;
+import roomescape.reservation.payload.ReservationRequest;
+import roomescape.reservation.payload.ReservationResponse;
+import roomescape.reservation.repository.ReservationRepository;
 
 @Controller
 public class ReservationController {
