+++ src/main/java/roomescape/reservation/entity/Reservation.java
@@ -1,4 +1,4 @@
-package roomescape.reservation;
+package roomescape.reservation.entity;
 
 import java.time.LocalDate;
