+++ src/main/java/roomescape/reservation/repository/JdbcReservationRepository.java
@@ -1,4 +1,4 @@
-package roomescape.reservation;
+package roomescape.reservation.repository;
 
 import java.sql.PreparedStatement;
 import java.time.LocalDate;
@@ -8,6 +8,8 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
+import roomescape.reservation.entity.Reservation;
+import roomescape.reservation.payload.ReservationRequest;
 
 @Repository
 public class JdbcReservationRepository implements ReservationRepository {
