+++ src/main/java/roomescape/reservation/repository/ReservationTimeRepository.java
@@ -1,6 +1,8 @@
-package roomescape.reservation;
+package roomescape.reservation.repository;
 
 import java.util.List;
+import roomescape.reservation.entity.ReservationTime;
+import roomescape.reservation.payload.ReservationTimeRequest;
 
 public interface ReservationTimeRepository {
